requirements.txt
absl-py==1.4.0
alabaster==0.7.13
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appnope==0.1.3
arabic-reshaper==2.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
async-generator==1.10
attrs==22.2.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.2
black==23.1.0
bleach==6.0.0
cachetools==4.2.4
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cloudpickle==2.2.1
comm==0.1.2
contourpy==1.0.7
curio==1.6
cycler==0.11.0
Cython==0.29.33
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@da8c5009b7f0c76519e4d97b19aca60406055fe3
diffimg==0.2.3
docrepr==0.2.0
docutils==0.18.1
entrypoints==0.4
exceptiongroup==1.1.0
executing==1.2.0
fastjsonschema==2.16.2
fire==0.5.0
fonttools==4.38.0
fqdn==1.5.1
future==0.18.3
fvcore @ git+https://github.com/facebookresearch/fvcore.git@2b14b2a025e60b0371a8509c42a2d45c821211c0
gast==0.3.3
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==2.10.0
hydra-core==1.3.1
idna==3.4
imageio==2.25.0
imagesize==1.4.1
imgaug==0.4.0
importlib-metadata==6.0.0
importlib-resources==5.10.2
iniconfig==2.0.0
iopath==0.1.9
ipykernel==6.21.1
ipyparallel==8.4.1
ipython==8.10.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==8.0.2
jupyter_core==5.2.0
jupyter_server==2.2.1
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
lmdb==1.4.0
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
mistune==2.0.5
mypy-extensions==1.0.0
natsort==8.2.0
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.0
nltk==3.8.1
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.21.0
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.7.0.68
opt-einsum==3.3.0
outcome==1.2.0
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
platformdirs==3.0.0
pluggy==1.0.0
portalocker==2.7.0
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==3.20.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
pycparser==2.21
Pygments==2.14.0
pyparsing==3.0.9
pyrsistent==0.19.3
pytest==6.2.5
pytest-asyncio==0.20.3
python-bidi==0.4.2
python-dateutil==2.8.2
python-json-logger==2.0.4
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
scikit-image==0.16.2
scipy==1.4.1
Send2Trash==1.8.0
shapely==2.0.1
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
Sphinx==6.1.3
sphinx-rtd-theme==1.2.0
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jquery==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stack-data==0.6.2
tabulate==0.9.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==2.2.0
terminado==0.17.1
testpath==0.6.0
tifffile==2023.2.3
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
torch==1.8.0
torchvision==0.9.0
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
trdg==1.8.0
trio==0.22.0
typing_extensions==4.4.0
uri-template==1.2.0
urllib3==1.26.14
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.2
widgetsnbextension==4.0.5
wikipedia==1.4.0
wrapt==1.14.1
yacs==0.1.8
zipp==3.13.0
